Move preferences stuff to gui.preferences
