+++ src/main/java/org/jabref/gui/actions/ShowPreferencesAction.java
@@ -1,25 +1,18 @@
 package org.jabref.gui.actions;
 
 import org.jabref.gui.JabRefFrame;
-import org.jabref.gui.preftabs.PreferencesDialog;
+import org.jabref.gui.preferences.PreferencesDialog;
 
 public class ShowPreferencesAction extends SimpleCommand {
 
-    private PreferencesDialog prefsDialog;
     private final JabRefFrame jabRefFrame;
-
     public ShowPreferencesAction(JabRefFrame jabRefFrame) {
         this.jabRefFrame = jabRefFrame;
     }
 
     @Override
     public void execute() {
-        if (prefsDialog == null) {
-            prefsDialog = new PreferencesDialog(jabRefFrame);
-        } else {
-            prefsDialog.setValues();
-        }
-
-        prefsDialog.showAndWait();
+        PreferencesDialog preferencesDialog = new PreferencesDialog(jabRefFrame);
+        preferencesDialog.showAndWait();
     }
 }
