+++ src/main/java/org/jabref/gui/bibtexkeypattern/BibtexKeyPatternPanel.java
@@ -15,7 +15,7 @@ import javafx.scene.layout.Pane;
 import org.jabref.Globals;
 import org.jabref.gui.BasePanel;
 import org.jabref.gui.help.HelpAction;
-import org.jabref.gui.preftabs.FontSize;
+import org.jabref.gui.preferences.FontSize;
 import org.jabref.logic.help.HelpFile;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.EntryTypes;
