+++ src/main/java/org/jabref/gui/preferences/ExternalTab.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.preftabs;
+package org.jabref.gui.preferences;
 
 import java.awt.BorderLayout;
