+++ src/main/java/org/jabref/gui/preferences/ImportSettingsTab.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.preftabs;
+package org.jabref.gui.preferences;
 
 import java.util.Objects;
