+++ src/main/java/org/jabref/gui/preferences/TablePrefsTab.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.preftabs;
+package org.jabref.gui.preferences;
 
 import java.util.Collections;
 import java.util.List;
